requirement.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
decorator==4.3.0
enum34==1.1.6
idna==2.8
ipython==5.6.0
ipython-genutils==0.2.0
lxml==4.3.0
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.3.1
pyquery==1.4.0
-e git://github.com/brokenseal/PyV8-OS-X@c91d6c54aea729161567c1229b12c66f62474513#egg=PyV8
requests==2.21.0
scandir==1.9.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
